Change back to using double instead of float

+++ PeriodicTopicBehaviors/src/main/java/edu/ntnu/app/psta/LatentWordByTL.java
@@ -4,7 +4,7 @@ import java.util.Arrays;
 
 public class LatentWordByTL implements Variable {
 
-    private final Float[][] latentWordByTL;
+    private final double[][] latentWordByTL;
     private final VariableList topics;
     private final VariableList topicDistDocs;
     private final VariableList topicDistTLs;
@@ -15,10 +15,10 @@ public class LatentWordByTL implements Variable {
         this.topicDistDocs = topicDistDocs;
         this.topicDistTLs = topicDistTLs;
         this.docIndex = docIndex;
-        this.latentWordByTL = new Float[PstaDocs.nWords()][topics.length()];
+        this.latentWordByTL = new double[PstaDocs.nWords()][topics.length()];
         for (int i = 0; i < PstaDocs.nWords(); i++) {
             // No need for initial values as we update the latent variables first. Must set to 0 as default is null...
-            this.latentWordByTL[i] = new Float[topics.length()];
+            this.latentWordByTL[i] = new double[topics.length()];
             Arrays.fill(this.latentWordByTL[i], 0f);
         }
     }
@@ -37,13 +37,13 @@ public class LatentWordByTL implements Variable {
         for (int z = 0; z < topics.length(); z++) {
             for (int w = 0; w < PstaDocs.nWords(); w++) {
                 // The first part, p(w|z), is not a part of the paper, but w is not included at all in the formula..
-                //Float numerator = topics.get(z).get(w) * PSTA.LAMBDA_TL * topicDistTLs
-                Float numerator = Psta.LAMBDA_TL * topicDistTLs
+                //double numerator = topics.get(z).get(w) * PSTA.LAMBDA_TL * topicDistTLs
+                double numerator = Psta.LAMBDA_TL * topicDistTLs
                         .get(PstaDocs.get(docIndex).getLocationId())
                         .get(PstaDocs.get(docIndex).getTimestampId(), z);
-                Float denominator = (1 - Psta.LAMBDA_TL) * topicDistDocs.get(docIndex).get(z) + numerator;
-                Float oldVal = latentWordByTL[w][z];
-                Float newVal = denominator != 0 ? numerator / denominator : 0;
+                double denominator = (1 - Psta.LAMBDA_TL) * topicDistDocs.get(docIndex).get(z) + numerator;
+                double oldVal = latentWordByTL[w][z];
+                double newVal = denominator != 0 ? numerator / denominator : 0;
                 converges = converges && Math.abs(oldVal - newVal) < Psta.EPSILON;
                 latentWordByTL[w][z] = newVal;
             }
@@ -57,7 +57,7 @@ public class LatentWordByTL implements Variable {
     }
 
     @Override
-    public Float get(int... values) {
+    public double get(int... values) {
         if (values.length != 2) {
             throw new IllegalArgumentException("Wrong number of values passed to LatentWordByTopic.get(). It should be 2.");
         }
